--- plugins/comparative-adapters/src/MCScanAnchorsAdapter/index.js
@@ -1,2 +1,0 @@
-export { default as AdapterClass } from './MCScanAnchorsAdapter'
-export { default as configSchema } from './configSchema'
