+++ plugins/comparative-adapters/src/index.ts
@@ -3,6 +3,7 @@ import PluginManager from '@jbrowse/core/PluginManager'
 import { FileLocation } from '@jbrowse/core/util/types'
 
 import PAFAdapterF from './PAFAdapter'
+import MCScanAnchorsAdapterF from './PAFAdapter'
 import DeltaAdapterF from './DeltaAdapter'
 import ChainAdapterF from './ChainAdapter'
 
@@ -19,6 +20,7 @@ export default class ComparativeAdaptersPlugin extends Plugin {
     PAFAdapterF(pluginManager)
     DeltaAdapterF(pluginManager)
     ChainAdapterF(pluginManager)
+    MCScanAnchorsAdapterF(pluginManager)
 
     pluginManager.addToExtensionPoint(
       'Core-guessAdapterForLocation',
